feat: functional end-to-end testing environment
This introduces a complete and functional testing environment for
testing the new activity service in a local kind cluster. The test
environment is built on top of the [test-infra] cluster that includes
base services like flux, envoy gateway, and a telemetry stack.

The test environment includes a Vector, NATS, Clickhouse pipeline that
automatically collects audit logs emitted from the test-infra kind
cluster.

The apiserver deployment manifests are structured as a standard `base`
kustomize deployment that includes the kubernetes Deployment, Service,
and RBAC resources.

The following kustomize components have been introduced to provide
optional functionality that can be enabled in environments when
necessary.

- api-registration: Configures the APIService registration with the k8s
  apiserver to proxy requests to the activity apiserver
- cert-manager-ca: Configures a namespaced cert issuer to use with the
  activity apiserver
- grafana-clickhouse: Configures a new Grafana datasource to connect to
  the deployed clickhouse instance
- namespace: Creates a namespace to use for the system's deployment
- nats-stream: Creates a new nats JetStream to use for the audit log
  pipeline
- tracing: Configures the APIserver with a tracing configuration
- vector-aggregator: Deploys an aggregated version of Vector that
  ingests audit logs from NATS and writes them to Clickhouse
- vector-sidecar: Deployment of Vector that runs on every node in the
  cluster that's responsible for collecting audit logs from apiservers
  and writing them to NATS

Also included is deployment automation for deploying the system's
dependencies:

- Clickhouse operator: Manages deployments of Clickhouse through CRDs
- NATS: Deploys an instance of NATS and NACK to configure NATS through
  CRDs

[test-infra]: https://github.com/datum-cloud/test-infra