feat: introduce the activity CLI
The Activity CLI is available for consumers to easily build CLIs that
can interact with the Activity API. The activity CLI includes the
following commands:

- `activity` - Entrypoint to all commands available with the activity
  CLI
- `query` - Command to execute queries against the API

This gives users the option of how they want to integrate the CLI and
provide a native experience to their users. CLIs can choose to import
just the `query` command if they wish.

This also includes a [kubectl-plugin] that makes it simple for users to
quickly get started with running the activity CLI natively with kubectl
as if it was a kubectl command.

[kubectl-plugin]:
    https://kubernetes.io/docs/tasks/extend-kubectl/kubectl-plugins/