feat: publish docker image and kustomize bundles
Creates a build pipeline that generates a docker image with the activity
command and all of the deployment manifests for the system. The
deployment manifests will only be published if the container image is
published.

This also introduces a base Taskfile command for building the service
and deploying the test infrastructure with a single command.